Ai_GUI/requirements.txt
appJar
Pillow